pyproject.toml
[tool.poetry]
name = "wikipedia-rag"
package-mode = false
version = "3.0.0"
description = "Real-time Wikipedia RAG with Groq LLaMA 3.3, Cohere Embeddings, and Pinecone"
authors = ["You <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
langchain = "^0.3.0"
langchain-groq = "^0.2.0"
langchain-cohere = "^0.3.0"
langchain-pinecone = "^0.2.0"
langchain-community = "^0.3.0"
python-dotenv = "^1.0.0"
requests = "^2.32.0"
pydantic = "^2.9.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"
notebook = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"